ccwl: Support arbitrary expressions in #:other arguments.
Support arbitrary expressions, that are actually evaluated, in #:other
arguments. This means, among other things, that users will have to
quote them correctly. While this complicates matters for novice users
who may not be familiar with scheme, it is the most general and most
flexible solution. In this particular case, we value flexibility over
ease of use since #:other is an escape hatch and if users have to use
it, ease of use has already been lost.

* ccwl/ccwl.scm (input, output, command): Make #:other an actually
evaluated unary argument.