ccwl: Restrict #:other parameter to YAML serializable trees.
Perhaps it fits in better with the spirit of ccwl to not evaluate
#:other parameters and instead allow only YAML serializable
trees. This saves users from quoting #:other parameters, of
course. But it also permits more thorough analysis and clearer error
messages. Besides, there is not much reason to allow general
expressions in #:other parameters. Such dynamism is not commonly
required in ccwl or CWL.

* ccwl/ccwl.scm (input, output, command): Restrict #:other parameter
to YAML serializable trees that are not evaluated.