build-aux: Do not infer release information from git tags.
We now build the website as a computed-file G-expression in a Guix
channel. Guix channels always discard the git repository information,
and don't have access to the git tags. We can fix this problem later
if and when we switch to using tissue for the website.

* build-aux/build-home-page.el (org-dblock-write:download): Do not
infer release information from git tags. Hard-code them.